+++ src/models/ComponentAttributeDefinition.js
@@ -10,6 +10,9 @@ class ComponentAttributeDefinition {
    * valid types are String/Boolean/Number/Array/Object/Link/Reference.
    * @param {String} [props.linkType] - Type of link, valid types are Default/Reverse.
    * @param {String[]} [props.linkRef=[]] - Reference of accepted component for link.
+   * @param {String} [props.linkColor='black'] - Color of the link.
+   * @param {Number} [props.linkWidth=2] - Width of the link.
+   * @param {Number[]} [props.linkDashStyle] - Dash style of the link. See stroke-dasharray of svg.
    * @param {String[]} [props.containerRef=[]] - Reference of accepted component for container.
    * @param {Boolean} [props.required=false] - Attribute is required.
    * @param {ComponentAttributeDefinition[]} [props.definedAttributes=[]] - Defined attributes for
@@ -25,6 +28,9 @@ class ComponentAttributeDefinition {
     type: null,
     linkType: null,
     linkRef: [],
+    linkColor: 'black',
+    linkWidth: 2,
+    linkDashStyle: null,
     containerRef: [],
     required: false,
     definedAttributes: [],
@@ -40,6 +46,9 @@ class ComponentAttributeDefinition {
       type,
       linkType,
       linkRef,
+      linkColor,
+      linkWidth,
+      linkDashStyle,
       containerRef,
       definedAttributes,
       required,
@@ -66,6 +75,21 @@ class ComponentAttributeDefinition {
      * @type {String[]}
      */
     this.linkRef = linkRef || [];
+    /**
+     * Color of the link.
+     * @type {String}
+     */
+    this.linkColor = linkColor || 'black';
+    /**
+     * Width of the link.
+     * @type {Number}
+     */
+    this.linkWidth = linkWidth || 2;
+    /**
+     * Dash style of the link. See stroke-dasharray of svg.
+     * @type {Number[]}
+     */
+    this.linkDashStyle = linkDashStyle || null;
     /**
      * Define list of Component that can be the container of this component.
      * @type {String[]}