+++ src/models/ComponentLinkDefinition.js
@@ -9,18 +9,27 @@ class ComponentLinkDefinition {
    * @param {String} [props.sourceRef] - Reference of component can be the source in a link.
    * @param {String} [props.targetRef] - Reference of component can be the target of the link.
    * @param {String} [props.type] - Representation of the link.
+   * @param {String} [props.color='black'] - Color of the link.
+   * @param {Number} [props.width=2] - Width of the link.
+   * @param {Number[]} [props.dashStyle] - Dash style of the link. See stroke-dasharray of svg.
    */
   constructor(props = {
     attributeRef: null,
     sourceRef: null,
     targetRef: null,
     type: null,
+    color: 'black',
+    width: 2,
+    dashStyle: null,
   }) {
     const {
       attributeRef,
       sourceRef,
       targetRef,
       type,
+      color,
+      width,
+      dashStyle,
     } = props;
 
     /**
@@ -43,6 +52,21 @@ class ComponentLinkDefinition {
      * @type {String}
      */
     this.type = type || null;
+    /**
+     * Color of the link.
+     * @type {String}
+     */
+    this.color = color || 'black';
+    /**
+     * Color of the link.
+     * @type {Number}
+     */
+    this.width = width || 2;
+    /**
+     * Dash style of the link. See stroke-dasharray of svg.
+     * @type {Number[]}
+     */
+    this.dashStyle = dashStyle || null;
   }
 }