+++ tests/unit/models/ComponentLinkDefinition.spec.js
@@ -9,6 +9,9 @@ describe('Test class: ComponentLinkDefinition', () => {
       expect(link.sourceRef).toBeNull();
       expect(link.targetRef).toBeNull();
       expect(link.type).toBeNull();
+      expect(link.color).toEqual('black');
+      expect(link.width).toEqual(2);
+      expect(link.dashStyle).toBeNull();
     });
 
     it('Check passing undefined variables to constructor', () => {
@@ -18,6 +21,9 @@ describe('Test class: ComponentLinkDefinition', () => {
       expect(link.sourceRef).toBeNull();
       expect(link.targetRef).toBeNull();
       expect(link.type).toBeNull();
+      expect(link.color).toEqual('black');
+      expect(link.width).toEqual(2);
+      expect(link.dashStyle).toBeNull();
     });
 
     it('Check passing variable to constructor', () => {
@@ -26,12 +32,18 @@ describe('Test class: ComponentLinkDefinition', () => {
         sourceRef: 'id1',
         targetRef: 'id2',
         type: 'type',
+        color: 'blue',
+        width: 3,
+        dashStyle: [1, 1],
       });
 
       expect(link.attributeRef).toEqual('attr');
       expect(link.sourceRef).toEqual('id1');
       expect(link.targetRef).toEqual('id2');
       expect(link.type).toEqual('type');
+      expect(link.color).toEqual('blue');
+      expect(link.width).toEqual(3);
+      expect(link.dashStyle).toEqual([1, 1]);
     });
   });
 });